Refactor things out of NPSP.py
